requirements.txt
Flask
Flask-CORS
pymongo
google-generativeai
python-dotenv
bcrypt
python-environ
flask_wtf